Split cli.rs into the commands it runs
One 3,565-line file held argument parsing, the health checks, install,
download, update, self-update, uninstall, serve, run submission, artifact
registration, listing, table printing, and a thousand lines of tests. Reading
any one of those meant scrolling past the other twelve.

It is now `adapters/cli/` with a module per command surface -- args, status,
install, update, uninstall, serve, run, show -- plus `shell` for the three
process helpers they share and `mod.rs` for the dispatcher and the prereq
gate. Every test moved next to what it tests; the one helper two modules need
(`run_args`, which parses an argv into RunArgs) is `test_support`, mirroring
`core/test_support.rs`.

Pure motion: no behaviour change, no test dropped, 142 still pass. Items are
`pub(super)` so the module boundaries are real -- `Command` stays private to
the dispatcher, and each command's helpers stay private to their module.
